refactoring: prep work extract to file + documentation
